organizing release data hooks
