+++ components/pages/about/clinical/ReleaseData.tsx
@@ -29,8 +29,8 @@ import Loader from '#components/Loader';
 import defaultTheme from '#components/theme';
 import { CoronaVirus, CrossHairs, File, Storage } from '#components/theme/icons';
 import { getConfig } from '#global/config';
-import useReleaseData from '#global/hooks/clinical/useReleaseData';
-import { ReleaseDataProps } from '#global/hooks/clinical/useReleaseData/types';
+import useReleaseData from '#global/hooks/useReleaseData/clinical';
+import { ReleaseClinicalDataProps } from '#global/hooks/useReleaseData/types';
 import useSingularityData from '#global/hooks/useSingularityData';
 
 const ReleaseData = (): ReactElement => {
@@ -40,7 +40,7 @@ const ReleaseData = (): ReactElement => {
 	const { fetchTotalCounts } = useSingularityData();
 
 	const [isLoadingSingularityData, setIsLoadingSingularityData] = useState<boolean>(true);
-	const [releaseDataProps, setReleaseDataProps] = useState<ReleaseDataProps>();
+	const [releaseDataProps, setReleaseDataProps] = useState<ReleaseClinicalDataProps>();
 
 	useEffect(() => {
 		fetchTotalCounts()
