+++ components/pages/clinical/DataAnalysis.tsx
@@ -27,7 +27,7 @@ import { Col, Row } from 'react-grid-system';
 
 import Loader from '#components/Loader';
 import { CoronaVirus, CrossHairs, File, Storage } from '#components/theme/icons';
-import useReleaseData from '#global/hooks/clinical/useReleaseData';
+import useReleaseData from '#global/hooks/useReleaseData/clinical';
 
 const DataAnalysis = (): ReactElement => {
 	const [releaseData, isFetchingData] = useReleaseData();
